wallet-api: Add blockchain index
* Emulator keeps track of all transactions on the blockchain
  because we will need to look up transactions by their hashes
  for validation
